move document page component
