+++ frontend/src/app/document/document-page/document-page.component.spec.ts
@@ -1,8 +1,8 @@
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-import { commonTestBed } from '../common-test-bed';
+import { commonTestBed } from '../../common-test-bed';
 
 import { DocumentPageComponent } from './document-page.component';
-import { makeDocument } from '../../mock-data/constructor-helpers';
+import { makeDocument } from '../../../mock-data/constructor-helpers';
 
 describe('DocumentPageComponent', () => {
     let component: DocumentPageComponent;
