+++ frontend/src/app/document/document-page/document-page.component.ts
